+++ src/components/Comparison/TableCell.js
@@ -1,13 +1,13 @@
 import { useState } from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import Tippy from '@tippyjs/react';
 import { ENTITIES } from 'constants/graphSettings';
 import { Button } from 'reactstrap';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import { faChevronCircleUp, faChevronCircleDown } from '@fortawesome/free-solid-svg-icons';
-import ValuePlugins from '../ValuePlugins/ValuePlugins';
-import StatementBrowserDialog from '../StatementBrowser/StatementBrowserDialog';
+import ValuePlugins from 'components/ValuePlugins/ValuePlugins';
+import StatementBrowserDialog from 'components/StatementBrowser/StatementBrowserDialog';
+import DescriptionTooltip from 'components/DescriptionTooltip/DescriptionTooltip';
 
 export const Item = styled.div`
     padding-right: 10px;
@@ -66,66 +66,68 @@ const TableCell = props => {
     };
 
     const PathTooltipContent = (data, cellDataValue) => (
-        <div className="fullPath">
-            Path of this value:{' '}
-            {data.pathLabels?.map((path, index) => {
-                const resourceType = isEqualPaths
-                    ? index % 2 === 0
+        <tr>
+            <td>Path</td>
+            <td>
+                {data.pathLabels?.map((path, index) => {
+                    const resourceType = isEqualPaths
+                        ? index % 2 === 0
+                            ? ENTITIES.RESOURCE
+                            : ENTITIES.PREDICATE
+                        : index % 2 !== 0
                         ? ENTITIES.RESOURCE
-                        : ENTITIES.PREDICATE
-                    : index % 2 !== 0
-                    ? ENTITIES.RESOURCE
-                    : ENTITIES.PREDICATE;
-                return (
-                    <span key={index}>
-                        <span
-                            className={resourceType !== ENTITIES.PREDICATE ? 'btn-link' : ''}
-                            onClick={() =>
-                                resourceType !== ENTITIES.PREDICATE
-                                    ? openStatementBrowser(
-                                          data.path[isEqualPaths ? index : index + 1],
-                                          path,
-                                          resourceType,
-                                          resourceType === ENTITIES.RESOURCE
-                                              ? data.pathLabels.slice(0, isEqualPaths ? index : index + 1).map((l, i) => ({
-                                                    id: cellDataValue.path[i],
-                                                    label: l,
-                                                    _class: i % 2 === 0 ? ENTITIES.RESOURCE : ENTITIES.PREDICATE,
-                                                }))
-                                              : [],
-                                      )
-                                    : null
-                            }
-                            style={{ cursor: resourceType !== ENTITIES.PREDICATE ? 'pointer' : 'default' }}
-                            onKeyDown={e =>
-                                e.keyCode === 13
-                                    ? () =>
-                                          resourceType !== ENTITIES.PREDICATE
-                                              ? openStatementBrowser(
-                                                    data.path[isEqualPaths ? index : index + 1],
-                                                    path,
-                                                    resourceType,
-                                                    resourceType === ENTITIES.RESOURCE
-                                                        ? data.pathLabels.slice(0, isEqualPaths ? index : index + 1).map((l, i) => ({
-                                                              id: cellDataValue.path[i],
-                                                              label: l,
-                                                              _class: i % 2 === 0 ? ENTITIES.RESOURCE : ENTITIES.PREDICATE,
-                                                          }))
-                                                        : [],
-                                                )
-                                              : null
-                                    : undefined
-                            }
-                            role="button"
-                            tabIndex={0}
-                        >
-                            {path}
+                        : ENTITIES.PREDICATE;
+                    return (
+                        <span key={index}>
+                            <span
+                                className={resourceType !== ENTITIES.PREDICATE ? 'btn-link' : ''}
+                                onClick={() =>
+                                    resourceType !== ENTITIES.PREDICATE
+                                        ? openStatementBrowser(
+                                              data.path[isEqualPaths ? index : index + 1],
+                                              path,
+                                              resourceType,
+                                              resourceType === ENTITIES.RESOURCE
+                                                  ? data.pathLabels.slice(0, isEqualPaths ? index : index + 1).map((l, i) => ({
+                                                        id: cellDataValue.path[i],
+                                                        label: l,
+                                                        _class: i % 2 === 0 ? ENTITIES.RESOURCE : ENTITIES.PREDICATE,
+                                                    }))
+                                                  : [],
+                                          )
+                                        : null
+                                }
+                                style={{ cursor: resourceType !== ENTITIES.PREDICATE ? 'pointer' : 'default' }}
+                                onKeyDown={e =>
+                                    e.keyCode === 13
+                                        ? () =>
+                                              resourceType !== ENTITIES.PREDICATE
+                                                  ? openStatementBrowser(
+                                                        data.path[isEqualPaths ? index : index + 1],
+                                                        path,
+                                                        resourceType,
+                                                        resourceType === ENTITIES.RESOURCE
+                                                            ? data.pathLabels.slice(0, isEqualPaths ? index : index + 1).map((l, i) => ({
+                                                                  id: cellDataValue.path[i],
+                                                                  label: l,
+                                                                  _class: i % 2 === 0 ? ENTITIES.RESOURCE : ENTITIES.PREDICATE,
+                                                              }))
+                                                            : [],
+                                                    )
+                                                  : null
+                                        : undefined
+                                }
+                                role="button"
+                                tabIndex={0}
+                            >
+                                {path}
+                            </span>
+                            {index !== data.pathLabels?.length - 1 && ' / '}
                         </span>
-                        {index !== data.pathLabels?.length - 1 && ' / '}
-                    </span>
-                );
-            })}
-        </div>
+                    );
+                })}
+            </td>
+        </tr>
     );
 
     const onClickHandle = (date, index) => {
@@ -162,12 +164,11 @@ const TableCell = props => {
                                 (date.type === ENTITIES.RESOURCE ? (
                                     <span key={`value-${date.resourceId}`}>
                                         {index > 0 && <ItemInnerSeparator />}
-                                        <Tippy
-                                            content={PathTooltipContent(date, data[index])}
-                                            arrow={true}
-                                            disabled={date.pathLabels?.length <= 1}
-                                            interactive={true}
-                                            appendTo={document.body}
+                                        <DescriptionTooltip
+                                            id={date.resourceId}
+                                            _class={date.type}
+                                            classes={date.classes ?? []}
+                                            extraContent={date.pathLabels?.length > 1 ? PathTooltipContent(date, data[index]) : ''}
                                         >
                                             <div
                                                 className="btn-link"
@@ -179,24 +180,22 @@ const TableCell = props => {
                                             >
                                                 <ValuePlugins type="resource">{date.label}</ValuePlugins>
                                             </div>
-                                        </Tippy>
+                                        </DescriptionTooltip>
                                     </span>
                                 ) : (
                                     <span key={`value-${date.label}-${index}`}>
                                         {index > 0 && <ItemInnerSeparator />}
-                                        <Tippy
-                                            content={PathTooltipContent(date, data[index])}
-                                            arrow={true}
-                                            disabled={date.pathLabels?.length <= 1}
-                                            interactive={true}
-                                            appendTo={document.body}
+                                        <DescriptionTooltip
+                                            id={date.resourceId}
+                                            _class={date.type}
+                                            extraContent={date.pathLabels?.length > 1 ? PathTooltipContent(date, data[index]) : ''}
                                         >
                                             <span>
                                                 <ValuePlugins type="literal" options={{ inModal: true }}>
                                                     {date.label}
                                                 </ValuePlugins>
                                             </span>
-                                        </Tippy>
+                                        </DescriptionTooltip>
                                     </span>
                                 )),
                         )}